+++ src/Commands/Make/ExceptionMakeCommand.php
@@ -9,21 +9,21 @@ use Nwidart\Modules\Traits\ModuleCommandTrait;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 
-class ServiceMakeCommand extends GeneratorCommand
+class ExceptionMakeCommand extends GeneratorCommand
 {
     use ModuleCommandTrait;
 
     protected $argumentName = 'name';
-    protected $name = 'module:make-service';
-    protected $description = 'Create a new service class for the specified module.';
+    protected $name = 'module:make-exception';
+    protected $description = 'Create a new exception class for the specified module.';
 
     public function getDestinationFilePath(): string
     {
         $path = $this->laravel['modules']->getModulePath($this->getModuleName());
 
-        $filePath = GenerateConfigReader::read('services')->getPath() ?? config('modules.paths.app_folder') . 'Services';
+        $filePath = GenerateConfigReader::read('exceptions')->getPath() ?? config('modules.paths.app_folder') . 'Exceptions';
 
-        return $path . $filePath . '/' . $this->getServiceName() . '.php';
+        return $path . $filePath . '/' . $this->getExceptionName() . '.php';
     }
 
     protected function getTemplateContents(): string
@@ -39,7 +39,7 @@ class ServiceMakeCommand extends GeneratorCommand
     protected function getArguments(): array
     {
         return [
-            ['name', InputArgument::REQUIRED, 'The name of the service class.'],
+            ['name', InputArgument::REQUIRED, 'The name of the action class.'],
             ['module', InputArgument::OPTIONAL, 'The name of module will be used.'],
         ];
     }
@@ -50,28 +50,37 @@ class ServiceMakeCommand extends GeneratorCommand
     protected function getOptions(): array
     {
         return [
-            ['invokable', 'i', InputOption::VALUE_NONE, 'Generate an invokable class', null],
+            ['render', '', InputOption::VALUE_NONE, 'Create the exception with an empty render method', null],
+            ['report', '', InputOption::VALUE_NONE, 'Create the exception with an empty report method', null],
             ['force', 'f', InputOption::VALUE_NONE, 'su.'],
         ];
     }
 
-    protected function getServiceName(): array|string
+    protected function getExceptionName(): array|string
     {
         return Str::studly($this->argument('name'));
     }
 
     private function getClassNameWithoutNamespace(): array|string
     {
-        return class_basename($this->getServiceName());
+        return class_basename($this->getExceptionName());
     }
 
     public function getDefaultNamespace(): string
     {
-        return config('modules.paths.generator.services.namespace', 'Services');
+        return config('modules.paths.generator.exceptions.namespace', 'Exceptions');
     }
 
     protected function getStubName(): string
     {
-        return $this->option('invokable') === true ? '/service-invoke.stub' : '/service.stub';
+        if ($this->option('render')) {
+            return $this->option('report')
+                ? '/exception-render-report.stub'
+                : '/exception-render.stub';
+        }
+
+        return $this->option('report')
+            ? '/exception-report.stub'
+            : '/exception.stub';
     }
 }